requirements.txt
langchain-core==0.3.40
langchain-openai==0.3.0
langgraph==0.2.67
openai==1.59.6
streamlit==1.43.2
python-dotenv
IPython==8.12.3
langchain
pydub
langchain-community
fastapi
uvicorn
ffmpeg
reportlab
langchain_groq
matplotlib
plotly
ldap3
msal
pyjwt
requests
pyodbc==5.2.0
azure-storage-blob
azure-identity
azure-keyvault-secrets
python-multipart
websockets
azure-cognitiveservices-speech
mutagen